requirements.txt
apns2==0.7.2
apple-music-python==1.0.5
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==2.0.0
boto3==1.26.0
botocore==1.29.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.2
django-cache-url==3.4.4
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-push-notifications==3.0.0
django-redis==5.2.0
django-rest-framework==0.1.0
django-storages==1.13.1
djangorestframework==3.14.0
environs==9.5.0
gunicorn==20.1.0
h2==2.6.2
hiredis==2.2.3
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
jmespath==1.0.1
marshmallow==3.19.0
msgpack==1.0.4
packaging==21.3
phonenumbers==8.13.15
Pillow==9.2.0
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydub==0.25.1
PyJWT==1.7.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
spotipy==2.21.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.5.0
